Added in 0.81.7 int06

+++ src/IDDBufferedOutputStream.java
@@ -121,16 +121,19 @@ public class IDDBufferedOutputStream extends BufferedOutputStream {
 		type=3;
 		IddOut=new BufferedOutputStream(new FileOutputStream(name,filenumber==1?false:true),655350);
 	}
+
 	public String renameVdrTo(String parent, String oldName){
 		String num = "000"+filenumber+".vdr";
 		String newName = parent + num.substring(num.length()-7);
-		File nname = new File(newName);
 
-		if (nname.exists())
+		File nname = new File(newName);
+		File oname = new File(oldName);
+
+		//DM09072004 081.7 int06 changed
+		if (!oname.getName().equals(nname.getName()) && nname.exists())
 			nname.delete();
 
-		Common.renameTo(new File(oldName), nname); //DM13042004 081.7 int01 changed
-		//new File(oldName).renameTo(nname);
+		Common.renameTo(oname, nname); //DM13042004 081.7 int01 changed
 
 		return newName;
 	}
